refactor: move code around w/out change to act as base for comparing
Moved the contract facet and it's jsdoc into a new file (mostly
without changes) to ease comparisons. The helper functions that
already existed in close to their final form were also moved to their
final locations for the same reason.

This version isn't intended to compile or run. All the corrections are
in the following commit.